Splits DMASTAggregate<> into two classes
Splits DMASTAggregate<> into two

One is for ProcStatements and the other is for TopStatements. This is necessary to allow for DMASTStatements to allow make use of aggregation (since they apparently are a separate branch of classes from DMASTProcStatement)

Both genres of statement need this to handle multiple declarations happening on the same path.

I wanted to preserve it being a single type, but after trying every weird type-engine hack I can think of, I'm just giving up. This is sadly the simplest thing to do.

Includes also some novel doc comments to make sure nobody else gets confused by this, like I did. :^)